backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
firebase-admin>=6.5.0
google-cloud-vision>=3.7.2
google-generativeai>=0.5.4
scikit-learn>=1.4.2
pandas>=2.2.2
numpy>=1.26.4
python-dateutil>=2.9.0
python-multipart>=0.0.9
pydantic>=2.7.1
httpx>=0.27.0
